cms/requirements.txt
streamlit
requests
cloudinary
python-dotenv
